+++ command/init.go
@@ -265,7 +265,7 @@ job "example" {
       #   }
       # }
 
-      # The "logs" stana instructs the Nomad client on how many log files and
+      # The "logs" stanza instructs the Nomad client on how many log files and
       # the maximum size of those logs files to retain. Logging is enabled by
       # default, but the "logs" stanza allows for finer-grained control over
       # the log rotation and storage configuration.